feature: added class DateRangeFilter
